backend/requirements.txt
fastapi==0.115.8
uvicorn[standard]==0.34.0
SQLAlchemy==2.0.38
psycopg2-binary==2.9.10
alembic==1.14.1
python-jose==3.3.0
passlib[bcrypt]==1.7.4
pydantic-settings==2.7.1
python-multipart==0.0.20
jinja2==3.1.6
